+++ tests/entity/test_integration.py
@@ -20,7 +20,7 @@ class TestEntityBaseTestCase(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super(TestEntityBaseTestCase, self).__init__(*args, **kwargs)
         uiza.workspace_api_domain = 'test domain'
-        uiza.api_key = 'test api key'
+        uiza.authorization = 'test api key'
         self.entity_id = '37d6706e-be91-463e-b3b3-b69451dd4752'
         self.entity_data_create = {
             "name": "Sample Video Python",