+++ tests/live/test_integration.py
@@ -21,7 +21,7 @@ class TestLiveBaseTestCase(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super(TestLiveBaseTestCase, self).__init__(*args, **kwargs)
         uiza.workspace_api_domain = 'test domain'
-        uiza.api_key = 'test api key'
+        uiza.authorization = 'test api key'
         self.live_id = '37d6706e-be91-463e-b3b3-b69451dd4752'
         self.live_data_create = {
             'name': 'test event python 1',
@@ -255,16 +255,6 @@ class TestListLive(TestLiveBaseTestCase):
         mock_request_http.return_value = True, 200
         with self.assertRaises(ClientException) as context:
             Live().list()
-        self.assertTrue(context.exception.__class__.__name__, 'ClientException')
-
-
-class TestDeleteLive(TestLiveBaseTestCase):
-
-    @mock.patch('uiza.Connection._request_http')
-    def test_delete_live_invalid(self, mock_request_http):
-        mock_request_http.return_value = True, 200
-        with self.assertRaises(ClientException) as context:
-            Live().delete(id='test id')
         self.assertTrue(context.exception.__class__.__name__, 'ClientException')
 
 
@@ -626,68 +616,68 @@ class TestDeleteRecordedLive(TestLiveBaseTestCase):
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_valid(self, mock_request_http):
         mock_request_http.return_value = True, 200
-        data = Live().delete_recorded(id='Test id recorded')
+        data = Live().delete(id='Test id recorded')
         self.assertEqual(data[1], 200)
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_400(self, mock_request_http):
         mock_request_http.return_value = True, 400
         with self.assertRaises(BadRequestError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'BadRequestError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_401(self, mock_request_http):
         mock_request_http.return_value = True, 401
         with self.assertRaises(UnauthorizedError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'UnauthorizedError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_404(self, mock_request_http):
         mock_request_http.return_value = True, 404
         with self.assertRaises(NotFoundError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'NotFoundError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_422(self, mock_request_http):
         mock_request_http.return_value = True, 422
         with self.assertRaises(UnprocessableError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'UnprocessableError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_500(self, mock_request_http):
         mock_request_http.return_value = True, 500
         with self.assertRaises(InternalServerError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'InternalServerError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_503(self, mock_request_http):
         mock_request_http.return_value = True, 503
         with self.assertRaises(ServiceUnavailableError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'ServiceUnavailableError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_4xx(self, mock_request_http):
         mock_request_http.return_value = True, 412
         with self.assertRaises(ClientError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'ClientError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_status_code_5xx(self, mock_request_http):
         mock_request_http.return_value = True, 512
         with self.assertRaises(ServerError) as context:
-            Live().delete_recorded(id='Test id recorded')
+            Live().delete(id='Test id recorded')
         self.assertTrue(context.exception.__class__.__name__, 'ServerError')
 
     @mock.patch('uiza.Connection._request_http')
     def test_delete_recorded_live_invalid_with_not_live_id(self, mock_request_http):
         mock_request_http.return_value = True, 200
         with self.assertRaises(Exception) as context:
-            Live().delete_recorded()
+            Live().delete()
         self.assertTrue(context.exception.__class__.__name__, 'TypeError')