+++ uiza/api_resources/category/category.py
@@ -8,7 +8,7 @@ from uiza.utility.utility import set_url
 class Category(UizaBase):
 
     def __init__(self):
-        self.connection = Connection(workspace_api_domain=uiza.workspace_api_domain, api_key=uiza.api_key)
+        self.connection = Connection(workspace_api_domain=uiza.workspace_api_domain, api_key=uiza.authorization)
         self.connection.url = set_url(
             workspace_api_domain=self.connection.workspace_api_domain,
             api_type=settings.uiza_api.category.type,
@@ -50,6 +50,6 @@ class Category(UizaBase):
             entityId=entity_id,
             metadataIds=metadata_ids
         )
-        result = self.connection.post(data_body)
+        result = self.connection.delete(data_body)
 
         return result