+++ CHANGELOG.md
@@ -46,3 +46,22 @@
 
 ### Documentation
 - N/A
+
+
+## 1.1.1
+
+### Fixed
+- In category service:
+    - Fix delete relation category
+
+### Documentation
+- N/A
+
+
+## 1.1.2
+
+### Changed
+- Update docs
+
+### Documentation
+- N/A